ml-services/sentiment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
datasets==2.14.6
aiofiles==23.2.1